+++ src/actions/constants/web3.js
@@ -7,5 +7,4 @@ export const CREATE: 'web3__create' = 'web3__create';
 export const READY: 'web3__ready' = 'web3__ready';
 export const BLOCK_UPDATED: 'web3__block_updated' = 'web3__block_updated';
 export const GAS_PRICE_UPDATED: 'web3__gas_price_updated' = 'web3__gas_price_updated';
-export const PENDING_TX_RESOLVED: 'web3__pending_tx_resolved' = 'web3__pending_tx_resolved';
 export const DISCONNECT: 'web3__disconnect' = 'web3__disconnect';