+++ src/reducers/AccountsReducer.js
@@ -20,9 +20,12 @@ export type Account = {
     +addressPath: Array<number>;
     +address: string;
     balance: string;
+    availableBalance: string;
+    sequence: number;
     nonce: number;
     block: number;
     transactions: number;
+    empty: boolean;
 }
 
 export type State = Array<Account>;